+++ H3IotProject.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.12.35707.178
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IotProject.BlazorDashboard", "IotProject.BlazorDashboard\IotProject.BlazorDashboard.csproj", "{C6DE0C00-BF1E-4069-963C-F2004C6DB23A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotProject.BlazorDashboard", "IotProject.BlazorDashboard\IotProject.BlazorDashboard.csproj", "{C6DE0C00-BF1E-4069-963C-F2004C6DB23A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IotProject.API", "IotProject.API\IotProject.API.csproj", "{C4EBA0E6-BB83-4C0F-8BC6-9E29F0F68E9A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotProject.API", "IotProject.API\IotProject.API.csproj", "{C4EBA0E6-BB83-4C0F-8BC6-9E29F0F68E9A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Backend", "Backend", "{8891851C-25E8-43DF-BA46-25204174EC8A}"
 EndProject
@@ -13,13 +13,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Frontend", "Frontend", "{54
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{116017EA-95A3-46E0-95E5-2D2C2727DAD8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IotProject.Shared", "IotProject.Shared\IotProject.Shared.csproj", "{579DB689-52CE-876D-E4AB-3AFFE9C35555}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotProject.Shared", "IotProject.Shared\IotProject.Shared.csproj", "{579DB689-52CE-876D-E4AB-3AFFE9C35555}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Libraries", "Libraries", "{81C90B35-1320-4D2C-A2C2-094630D3E335}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IotProject.Auth", "IotProject.Auth\IotProject.Auth.csproj", "{77B1C480-B2F9-416B-9DC6-9650D61B89D8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotProject.Auth", "IotProject.Auth\IotProject.Auth.csproj", "{77B1C480-B2F9-416B-9DC6-9650D61B89D8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IotProject.Mobile", "IotProject.Mobile\IotProject.Mobile.csproj", "{5CFE38F9-8A57-4A75-BD2B-0713B0332F96}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IotProject.Mobile", "IotProject.Mobile\IotProject.Mobile.csproj", "{5CFE38F9-8A57-4A75-BD2B-0713B0332F96}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
