+++ IotProject.BlazorDashboard/Components/Layout/MainLayout.razor
@@ -38,6 +38,8 @@
                        new NavItem { Id = "1", Href = "/", IconName = IconName.HouseDoorFill, Text = "Home", Match=NavLinkMatch.All},
                        new NavItem { Id = "2", Href = "/counter", IconName = IconName.PlusSquareFill, Text = "Counter"},
                        new NavItem { Id = "3", Href = "/weather", IconName = IconName.Table, Text = "Fetch Data"},
+                       new NavItem { Id = "4", Href = "/login", IconName = IconName.Key, Text = "Login"},
+                       new NavItem { Id = "5", Href = "/logout", IconName = IconName.PersonLock, Text = "Logout"}
                    };
 
         return navItems;
