+++ IotProject.BlazorDashboard/Components/Pages/Auth/Logout.razor
@@ -1,0 +1,11 @@
+﻿@page "/logout"
+@inject NavigationManager NavigationManager
+@inject AuthService AuthService
+
+@code {
+	protected override async Task OnInitializedAsync()
+	{
+		await AuthService.Logout();
+		NavigationManager.NavigateTo("/", true);
+	}
+}
